Fix merge mistake #2
